refactor(renderer): extract pane-traversal helpers, dedupe findActiveLeaf
Move findSurfaceByPtyId + findActiveLeaf into src/renderer/utils/paneTraversal.ts as pure functions and drop the two inline findActiveLeaf closures in useNotificationListener. Behaviour is byte-for-byte identical; the R1-R4 notification-routing tests stay green.
